+++ flutter_plugin_qpos/CHANGELOG.md
@@ -1,4 +1,5 @@
-## 0.0.8
+## 0.0.9
 
-* 1. Add anlysEmvIccData implementation and updateWorkKey Implemented for IOS;
-* 2. fix sendOnlineProcessResult cannot process issue
+* 1. Add callbacks for bluetooth power on and power off on flutter.
+* 2. fix scan bluetooth twice issue(IOS).
+* 3. Add function for mifare operation: pollOnMifareCard; authenticateMifareCard; operateMifareCardData; readMifareCard; writeMifareCard; finishMifareCard